Add the --force option
I've been trying for 48 hours straight to get ionic to work. I've an old npm install with tons of cruft and kept getting error messages that no amount of googling could solve. Someone in the ionic forum suggested this command (which I'd tried before) with the force option, and it fixed my problem.